requirements.txt
google-genai>=1.66.0
google-cloud-texttospeech>=2.34.0
google-cloud-speech>=2.28.0
python-dotenv>=1.2.2
